+++ conftest.py
@@ -1,6 +1,0 @@
-import pytest
-
-
-@pytest.fixture(scope='session', autouse=True)
-def init_config():
-	__import__('pmxbot').config = {}