+++ pmxbot/testing/fixtures.py
@@ -1,0 +1,13 @@
+import pytest
+
+
+@pytest.fixture(scope='session', autouse=True)
+def init_config():
+	__import__('pmxbot').config = {}
+
+
+@pytest.fixture(params=['mongodb', 'sqlite'])
+def db_uri(request):
+	if request.param == 'mongodb':
+		return request.getfuncargvalue('mongodb_uri')
+	return 'sqlite:pmxbot.sqlite'